Generate Synthetic Data

Data_Generation generates synthetic data, where each covariate is a binary variable. The function used to create synthetic data can be found

Data_Generation(num_control, num_treated, num_cov_dense,
  num_cov_unimportant, U)

Arguments

num_control
number of samples in the control group

num_treated
number of samples in the treated group

num_cov_dense
number of important covariates

num_cov_unimportant
number of unimportant covariates

U
non-linear term coefficient

Value

synthetic data

Examples

Data_Generation(10, 10, 10, 5, 5)

#>    x1 x2 x3 x4 x5 x6 x7 x8 x9 x10 x11 x12 x13 x14 x15    outcome treated
#> 1   0  1  0  0  1  0  1  0  1   1   1   0   1   0   1  -5.051119       0
#> 2   1  0  1  0  0  0  0  0  1   0   0   1   1   1   0 -19.570606       0
#> 3   1  0  1  0  1  0  1  0  0   1   1   1   0   0   0 -13.706777       0
#> 4   0  0  0  1  1  1  0  0  0   1   1   1   1   0   0   1.096644       0
#> 5   0  0  1  0  0  0  1  0  0   1   0   1   0   1   1  -3.886992       0
#> 6   0  0  1  0  1  1  0  0  1   0   1   1   1   0   0 -15.102483       0
#> 7   0  0  0  1  1  0  1  0  0   1   0   0   1   1   0 -12.918575       0
#> 8   0  0  1  1  1  1  1  1  0   1   0   0   0   1   0 -17.789692       0
#> 9   1  0  1  0  1  0  1  1  1   0   1   0   1   1   1 -38.357507       0
#> 10  1  1  1  0  0  0  0  0  0   0   0   1   1   1   0  -1.531397       0
#> 11  1  1  1  0  0  1  1  1  1   1   1   1   0   0   0  20.280768       1
#> 12  1  1  1  1  1  0  1  0  0   0   0   1   1   0   1  35.481800       1
#> 13  0  0  0  1  1  1  1  0  1   0   1   1   1   0   0 -12.019617       1
#> 14  0  1  1  1  0  0  1  1  0   1   1   1   1   1   1  13.975389       1
#> 15  0  1  1  1  1  0  1  0  1   1   0   0   0   1   1  20.260162       1
#> 16  1  1  0  1  0  0  1  1  0   0   0   0   0   0   1   6.076932       1
#> 17  1  0  0  1  1  1  1  1  0   1   0   1   1   0   1  11.336012       1
#> 18  1  1  1  0  0  0  0  0  0   1   0   0   1   1   1  29.078950       1
#> 19  1  1  0  0  1  0  0  0  1   0   1   1   0   1   1  10.524320       1
#> 20  1  1  0  1  1  1  1  1  0   1   0   0   1   1   0  36.287564       1

Data_Generation(10, 10, 5, 10, 5)

#>    x1 x2 x3 x4 x5 x6 x7 x8 x9 x10 x11 x12 x13 x14 x15    outcome treated
#> 1   1  0  1  1  1  0  0  1  1   0   0   0   1   1   0  1.1776418       0
#> 2   1  1  0  1  0  1  1  1  0   0   0   0   1   1   1  1.5367338       0
#> 3   0  0  0  0  1  1  0  0  0   1   0   1   0   1   1 -2.9253989       0
#> 4   0  1  1  0  1  0  0  1  1   1   0   1   1   0   1 -1.4255353       0
#> 5   0  0  1  0  1  0  1  0  1   0   0   1   0   0   1 -0.8923137       0
#> 6   1  0  0  1  0  0  1  0  0   1   0   1   1   1   1  2.0699555       0
#> 7   0  1  1  0  1  1  1  0  0   1   1   0   1   0   0 -1.4255353       0
#> 8   0  1  1  1  1  1  1  0  1   0   1   0   1   0   0 -6.7667769       0
#> 9   1  1  1  1  1  0  1  1  0   0   1   1   1   1   0  0.6444201       0
#> 10  1  1  1  0  0  1  0  0  0   0   0   0   1   1   1  8.9110607       0
#> 11  0  1  0  1  0  1  0  1  1   0   1   0   1   0   1  2.2433408       1
#> 12  1  1  0  0  1  1  1  1  1   1   0   0   1   1   0 23.5724377       1
#> 13  1  0  0  1  0  1  1  0  0   1   1   0   0   0   1 10.0492393       1
#> 14  0  0  0  1  0  0  1  1  1   0   1   1   1   1   1 -3.8247280       1
#> 15  0  0  1  1  1  0  1  0  1   0   0   0   0   1   0 13.2494564       1
#> 16  0  0  0  1  0  0  0  0  0   0   1   1   0   1   1 -3.8247280       1
#> 17  0  0  0  0  1  1  1  1  1   1   1   1   1   1   1 -1.3695985       1
#> 18  0  0  1  1  1  0  0  0  1   0   1   1   1   0   0 13.2494564       1
#> 19  0  1  1  1  0  1  0  1  1   1   1   0   0   1   1 15.6871237       1
#> 20  0  1  1  1  0  0  1  1  1   1   0   1   1   1   0 15.6871237       1