SQLite Database Implementation

FLAME_SQLite applies the FLAME algorithm based on SQLite. FLAME_SQLite does not require external database installment. However, user should connect to a temporary database with command dbConnect(SQLite(),"tempdb_name") and name the connection as db. The required arguments include (1) db, (2) data, and (3) holdout. The rest of the arguments are optional.

FLAME_SQLite(db, data, holdout, compute_var = FALSE, tradeoff = 0.1,
  PE_function = NULL, model = NULL, ridge_reg = NULL,
  lasso_reg = NULL, tree_depth = NULL)

Arguments

db
name of the connection to temporary database (must name the connection as db)

data
input data

holdout
holdout training data

compute_var
indicator variable of computing variance (optional, default = FALSE)

tradeoff
tradeoff parameter to compute Matching Quality (optional, default = 0.1)

PE_function
user defined function to compute predivtive error (optional)

model
user defined model - Linear, Ridge, Lasso, or DecisionTree (optional)

ridge_reg
L2 regularization parameter if model = Ridge (optional)

lasso_reg
L1 regularization parameter if model = Lasso (optional)

tree_depth
maximum depth of decision tree if model = DecisionTree (optional)

Value

(1) list of covariates FLAME performs matching at each iteration, (2) list of dataframe showing matched groups' sizes, conditional average treatment effects (CATEs), and variance (if compute_var = TRUE) at each iteration, (3) matching quality at each iteration, and (4) the original data with additional column *matched*, indicating the number of covariates each unit is matched. If a unit is never matched, then *matched* will be 0.

Examples

# NOT RUN {
data <- data(toy_data)

db <- dbConnect(SQLite(),"tempdb")

FLAME_SQLite(db = db, data = data, holdout = data)

dbDisconnect(db)
# }